jit/requirements.txt
click
GitPython
setuptools